+++ test/Sdk.Generator.Tests/FunctionMetadataProviderGeneratorTests/DependentAssemblyTest.cs
@@ -67,11 +67,12 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                 using System.Collections.Immutable;
                 using System.Text.Json;
                 using System.Threading.Tasks;
+                using Microsoft.Azure.Functions.Worker;
                 using Microsoft.Azure.Functions.Worker.Core.FunctionMetadata;
                 using Microsoft.Extensions.DependencyInjection;
                 using Microsoft.Extensions.Hosting;
 
-                namespace Microsoft.Azure.Functions.Worker
+                namespace TestProject
                 {
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {