+++ test/Sdk.Generator.Tests/FunctionMetadataProviderGeneratorTests/IntegratedTriggersAndBindingsTests.cs
@@ -88,11 +88,12 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                 using System.Collections.Immutable;
                 using System.Text.Json;
                 using System.Threading.Tasks;
+                using Microsoft.Azure.Functions.Worker;
                 using Microsoft.Azure.Functions.Worker.Core.FunctionMetadata;
                 using Microsoft.Extensions.DependencyInjection;
                 using Microsoft.Extensions.Hosting;
 
-                namespace Microsoft.Azure.Functions.Worker
+                namespace TestProject
                 {
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
@@ -201,11 +202,12 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                 using System.Collections.Immutable;
                 using System.Text.Json;
                 using System.Threading.Tasks;
+                using Microsoft.Azure.Functions.Worker;
                 using Microsoft.Azure.Functions.Worker.Core.FunctionMetadata;
                 using Microsoft.Extensions.DependencyInjection;
                 using Microsoft.Extensions.Hosting;
 
-                namespace Microsoft.Azure.Functions.Worker
+                namespace TestProject
                 {
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
@@ -302,11 +304,12 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                 using System.Collections.Immutable;
                 using System.Text.Json;
                 using System.Threading.Tasks;
+                using Microsoft.Azure.Functions.Worker;
                 using Microsoft.Azure.Functions.Worker.Core.FunctionMetadata;
                 using Microsoft.Extensions.DependencyInjection;
                 using Microsoft.Extensions.Hosting;
 
-                namespace Microsoft.Azure.Functions.Worker
+                namespace TestProject
                 {
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
@@ -388,11 +391,12 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                 using System.Collections.Immutable;
                 using System.Text.Json;
                 using System.Threading.Tasks;
+                using Microsoft.Azure.Functions.Worker;
                 using Microsoft.Azure.Functions.Worker.Core.FunctionMetadata;
                 using Microsoft.Extensions.DependencyInjection;
                 using Microsoft.Extensions.Hosting;
 
-                namespace Microsoft.Azure.Functions.Worker
+                namespace TestProject
                 {
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
@@ -473,11 +477,12 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                 using System.Collections.Immutable;
                 using System.Text.Json;
                 using System.Threading.Tasks;
+                using Microsoft.Azure.Functions.Worker;
                 using Microsoft.Azure.Functions.Worker.Core.FunctionMetadata;
                 using Microsoft.Extensions.DependencyInjection;
                 using Microsoft.Extensions.Hosting;
 
-                namespace Microsoft.Azure.Functions.Worker
+                namespace TestProject
                 {
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
@@ -568,11 +573,12 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                 using System.Collections.Immutable;
                 using System.Text.Json;
                 using System.Threading.Tasks;
+                using Microsoft.Azure.Functions.Worker;
                 using Microsoft.Azure.Functions.Worker.Core.FunctionMetadata;
                 using Microsoft.Extensions.DependencyInjection;
                 using Microsoft.Extensions.Hosting;
 
-                namespace Microsoft.Azure.Functions.Worker
+                namespace TestProject
                 {
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {